requirements.txt
numpy==1.21.3
pandas==1.3.4
scikit-learn==1.0.1
scipy==1.7.1
hydra-core==1.1.1
--find-links https://download.pytorch.org/whl/cu102/torch_stable.html
torch==1.10.1+cu102
tensorboard==2.7.0
tqdm==4.62.3
